refactor: extract console-printer lib out of src folder
