+++ lib/console-printer/index.ts
