+++ src/reporting/reporters/internal/environment.ts
@@ -1,9 +1,9 @@
-import { EnvironmentContext } from "../../../config/environment";
-import { Nsci } from "../../../config/standard/index.js";
 import {
   consolePrinter,
   removeWhiteSpaces
-} from "../../../lib/console-printer/index.js";
+} from "../../../../lib/console-printer/index.js";
+import { EnvironmentContext } from "../../../config/environment";
+import { Nsci } from "../../../config/standard/index.js";
 import { Reporter } from "../reporter.js";
 
 function invertRecord(obj: Record<string, string>): Record<string, string> {
