+++ tsconfig.json
@@ -1,9 +1,7 @@
 {
   "compilerOptions": {
     "target": "es2020",
-    "lib": [
-      "es2020"
-    ],
+    "lib": ["es2020"],
     "module": "es2020",
     "moduleResolution": "Node",
     "esModuleInterop": true,
@@ -12,12 +10,6 @@
     "skipLibCheck": true,
     "allowSyntheticDefaultImports": true
   },
-  "include": [
-    "src",
-    "bin"
-  ],
-  "exclude": [
-    "node_modules",
-    "dist"
-  ]
+  "include": ["src", "bin", "lib"],
+  "exclude": ["node_modules", "dist"]
 }
